requirements.txt
yt-dlp>=2024.12.13
curl_cffi
ffmpeg-python
psycopg2-binary
requests                 